feat: moved vite config to one file
