+++ packages/solid-crs-components/package.json
@@ -18,8 +18,8 @@
   ],
   "scripts": {
     "start": "npm run start:watch",
-    "start:watch": "vite --config vite.serve.config.ts",
-    "build": "rimraf ./dist && vite build --config vite.build.config.ts && tsc",
+    "start:watch": "vite --config vite.config.ts",
+    "build": "rimraf ./dist && vite build --config vite.config.ts && tsc",
     "lint": "eslint .",
     "lint:fix": "eslint --fix .",
     "lint:staged": "lint-staged",
