--- packages/solid-crs-components/vite.build.config.ts
@@ -1,14 +1,0 @@
-import path from 'path';
-import { defineConfig } from 'vite'
-
-export default defineConfig({
-  root: 'lib',
-  build: {
-      target: 'es2015',
-      lib: {
-          entry: path.resolve(__dirname, 'lib/index.ts'),
-          name: '@netwerk-digitaal-erfgoed/solid-crs-components'
-      },
-      outDir: '../dist'
-  }
-});
