--- packages/solid-crs-components/vite.serve.config.ts
@@ -1,12 +1,0 @@
-import { defineConfig } from 'vite'
-
-export default defineConfig({
-  root: 'lib',
-  build: {
-      target: 'es2015',
-      outDir: '../dist'
-  },
-  server: {
-    port: 3001,
-  }
-});
