+++ packages/solid-crs-manage/package.json
@@ -18,8 +18,8 @@
   ],
   "scripts": {
     "start": "http-server -p 3002 --gzip --brotli --proxy 'http://localhost:3002?' dist/",
-    "start:watch": "vite serve --config vite.serve.config.ts",
-    "build": "rimraf ./dist && vite build --config vite.build.config.ts && tsc",
+    "start:watch": "vite serve --config vite.config.ts",
+    "build": "rimraf ./dist && vite build --config vite.config.ts && tsc",
     "lint": "eslint .",
     "lint:fix": "eslint --fix .",
     "lint:staged": "lint-staged",
