+++ packages/solid-crs-manage/vite.config.ts
@@ -1,0 +1,38 @@
+import { defineConfig, loadEnv } from 'vite';
+
+export default defineConfig( ({ command, mode }) => {
+  const env = loadEnv(mode, process.cwd())
+
+  // expose .env as process.env instead of import.meta since jest does not import meta yet
+  const envWithProcessPrefix = Object.entries(env).reduce(
+    (prev, [key, val]) => {
+      return {
+        ...prev,
+        ['process.env.' + key]: `"${val}"`,
+      }
+    },
+    {},
+  );
+
+  if (command === 'serve') {
+      return {
+        root: 'lib',
+        server: {
+          port: 3002,
+        },
+        define: envWithProcessPrefix,
+      }
+  } else if (command === 'build'){
+    return {
+        root: 'lib',
+        build: {
+            target: 'es2015',
+            outDir: '../dist',
+        },
+        server: {
+          port: 3002,
+        },
+        define: envWithProcessPrefix,
+      }
+  }
+});
